refactor: split i18n modules
